Move getFirewallDescription to Firewalls/shared.ts
